src/requirements.txt
pandas
sqlalchemy
python-dotenv
psycopg2-binary 
yfinance

